+++ easybill-java-sdk/src/main/java/io/github/bolzer/easybill_java_sdk/resources/ContactResource.java
@@ -1,7 +1,6 @@
 package io.github.bolzer.easybill_java_sdk.resources;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.github.bolzer.easybill_java_sdk.contracts.HttpClient;
 import io.github.bolzer.easybill_java_sdk.exceptions.EasybillRestException;
 import io.github.bolzer.easybill_java_sdk.models.Contact;
@@ -17,7 +16,6 @@ public final class ContactResource {
     public static final String RESOURCE_URL = "/customers";
 
     @NonNull
-    @SuppressFBWarnings
     private final HttpClient httpClient;
 
     public ContactResource(@NonNull HttpClient httpClient) {
@@ -28,7 +26,7 @@ public final class ContactResource {
         long customerId,
         GenericListQueryRequest genericListQueryRequest
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/" + customerId + "/contacts",
                 genericListQueryRequest,
                 new TypeReference<>() {}
@@ -37,7 +35,7 @@ public final class ContactResource {
 
     public @NonNull Contact fetchContact(long customerId, long contactId)
         throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/" + customerId + "/contacts/" + contactId,
                 new TypeReference<>() {}
             );
@@ -47,7 +45,7 @@ public final class ContactResource {
         long customerId,
         @NonNull ContactRequest contactRequest
     ) throws EasybillRestException {
-        return this.httpClient.postJson(
+        return this.httpClient.requestPostJson(
                 RESOURCE_URL + "/" + customerId + "/contacts",
                 contactRequest,
                 new TypeReference<>() {}
@@ -59,7 +57,7 @@ public final class ContactResource {
         @Positive long contactId,
         @NonNull ContactRequest contactRequest
     ) throws EasybillRestException {
-        return this.httpClient.putJson(
+        return this.httpClient.requestPutJson(
                 RESOURCE_URL + "/" + customerId + "/contacts/" + contactId,
                 contactRequest,
                 new TypeReference<>() {}
@@ -70,7 +68,7 @@ public final class ContactResource {
         @Positive long customerId,
         @Positive long contactId
     ) throws EasybillRestException {
-        this.httpClient.delete(
+        this.httpClient.requestDelete(
                 RESOURCE_URL + "/" + customerId + "/contacts/" + contactId
             );
     }