+++ easybill-java-sdk/src/main/java/io/github/bolzer/easybill_java_sdk/resources/DiscountResource.java
@@ -1,7 +1,6 @@
 package io.github.bolzer.easybill_java_sdk.resources;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.github.bolzer.easybill_java_sdk.contracts.HttpClient;
 import io.github.bolzer.easybill_java_sdk.exceptions.EasybillRestException;
 import io.github.bolzer.easybill_java_sdk.models.PositionDiscount;
@@ -19,7 +18,6 @@ public final class DiscountResource {
     public static final String RESOURCE_URL = "/discounts";
 
     @NonNull
-    @SuppressFBWarnings
     private final HttpClient httpClient;
 
     public DiscountResource(@NonNull HttpClient httpClient) {
@@ -29,7 +27,7 @@ public final class DiscountResource {
     public @NonNull PaginatedResponse<PositionDiscount> fetchPositionDiscounts(
         @NonNull DiscountListQueryRequest discountListQueryRequest
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/position",
                 discountListQueryRequest,
                 new TypeReference<>() {}
@@ -39,7 +37,7 @@ public final class DiscountResource {
     public @NonNull PositionDiscount createPositionDiscount(
         @NonNull PositionDiscountRequest positionDiscountRequest
     ) throws EasybillRestException {
-        return this.httpClient.postJson(
+        return this.httpClient.requestPostJson(
                 RESOURCE_URL + "/position",
                 positionDiscountRequest,
                 new TypeReference<>() {}
@@ -49,7 +47,7 @@ public final class DiscountResource {
     public @NonNull PositionDiscount fetchPositionDiscount(
         @Positive long positionDiscountId
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/position/" + positionDiscountId,
                 new TypeReference<>() {}
             );
@@ -59,7 +57,7 @@ public final class DiscountResource {
         @Positive long positionDiscountId,
         @NonNull PositionDiscountRequest positionDiscountRequest
     ) throws EasybillRestException {
-        return this.httpClient.putJson(
+        return this.httpClient.requestPutJson(
                 RESOURCE_URL + "/position/" + positionDiscountId,
                 positionDiscountRequest,
                 new TypeReference<>() {}
@@ -68,7 +66,7 @@ public final class DiscountResource {
 
     public void deletePositionDiscount(@Positive long positionDiscountId)
         throws EasybillRestException {
-        this.httpClient.delete(
+        this.httpClient.requestDelete(
                 RESOURCE_URL + "/position/" + positionDiscountId
             );
     }
@@ -76,7 +74,7 @@ public final class DiscountResource {
     public @NonNull PaginatedResponse<PositionGroupDiscount> fetchPositionGroupDiscounts(
         @NonNull DiscountListQueryRequest discountListQueryRequest
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/position-group",
                 discountListQueryRequest,
                 new TypeReference<>() {}
@@ -86,7 +84,7 @@ public final class DiscountResource {
     public @NonNull PositionGroupDiscount createPositionGroupDiscount(
         @NonNull PositionGroupDiscountRequest positionGroupDiscountRequest
     ) throws EasybillRestException {
-        return this.httpClient.postJson(
+        return this.httpClient.requestPostJson(
                 RESOURCE_URL + "/position-group",
                 positionGroupDiscountRequest,
                 new TypeReference<>() {}
@@ -96,7 +94,7 @@ public final class DiscountResource {
     public @NonNull PositionGroupDiscount fetchPositionGroupDiscount(
         @Positive long positionDiscountId
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/position-group/" + positionDiscountId,
                 new TypeReference<>() {}
             );
@@ -106,7 +104,7 @@ public final class DiscountResource {
         @Positive long positionDiscountId,
         @NonNull PositionGroupDiscountRequest positionGroupDiscountRequest
     ) throws EasybillRestException {
-        return this.httpClient.putJson(
+        return this.httpClient.requestPutJson(
                 RESOURCE_URL + "/position-group/" + positionDiscountId,
                 positionGroupDiscountRequest,
                 new TypeReference<>() {}
@@ -115,7 +113,7 @@ public final class DiscountResource {
 
     public void deletePositionGroupDiscount(@Positive long positionDiscountId)
         throws EasybillRestException {
-        this.httpClient.delete(
+        this.httpClient.requestDelete(
                 RESOURCE_URL + "/position-group/" + positionDiscountId
             );
     }