+++ easybill-java-sdk/src/main/java/io/github/bolzer/easybill_java_sdk/resources/PdfTemplateResource.java
@@ -1,7 +1,6 @@
 package io.github.bolzer.easybill_java_sdk.resources;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.github.bolzer.easybill_java_sdk.contracts.HttpClient;
 import io.github.bolzer.easybill_java_sdk.exceptions.EasybillRestException;
 import io.github.bolzer.easybill_java_sdk.models.PdfTemplate;
@@ -15,7 +14,6 @@ public final class PdfTemplateResource {
     public static final String RESOURCE_URL = "/pdf-templates";
 
     @NonNull
-    @SuppressFBWarnings
     private final HttpClient httpClient;
 
     public PdfTemplateResource(@NonNull HttpClient httpClient) {
@@ -24,13 +22,16 @@ public final class PdfTemplateResource {
 
     public @NonNull NonPaginatedResponse<PdfTemplate> fetchPdfTemplates()
         throws EasybillRestException {
-        return this.httpClient.getJson(RESOURCE_URL, new TypeReference<>() {});
+        return this.httpClient.requestGetJson(
+                RESOURCE_URL,
+                new TypeReference<>() {}
+            );
     }
 
     public @NonNull NonPaginatedResponse<PdfTemplate> fetchPdfTemplates(
         PdfTemplateRequest pdfTemplateRequest
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL,
                 pdfTemplateRequest,
                 new TypeReference<>() {}