+++ easybill-java-sdk/src/main/java/io/github/bolzer/easybill_java_sdk/resources/PostBoxResource.java
@@ -1,7 +1,6 @@
 package io.github.bolzer.easybill_java_sdk.resources;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.github.bolzer.easybill_java_sdk.contracts.HttpClient;
 import io.github.bolzer.easybill_java_sdk.exceptions.EasybillRestException;
 import io.github.bolzer.easybill_java_sdk.models.PostBox;
@@ -16,7 +15,6 @@ public final class PostBoxResource {
     public static final String RESOURCE_URL = "/post-boxes";
 
     @NonNull
-    @SuppressFBWarnings
     private final HttpClient httpClient;
 
     public PostBoxResource(@NonNull HttpClient httpClient) {
@@ -26,7 +24,7 @@ public final class PostBoxResource {
     public @NonNull PaginatedResponse<PostBox> fetchPostBoxes(
         PostBoxListQueryRequest postBoxListQueryRequest
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL,
                 postBoxListQueryRequest,
                 new TypeReference<>() {}
@@ -35,7 +33,7 @@ public final class PostBoxResource {
 
     public @NonNull PostBox fetchPostBox(@Positive long postBoxId)
         throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/" + postBoxId,
                 new TypeReference<>() {}
             );
@@ -43,6 +41,6 @@ public final class PostBoxResource {
 
     public void deletePostBox(@Positive long postBoxId)
         throws EasybillRestException {
-        this.httpClient.delete(RESOURCE_URL + "/" + postBoxId);
+        this.httpClient.requestDelete(RESOURCE_URL + "/" + postBoxId);
     }
 }