+++ easybill-java-sdk/src/main/java/io/github/bolzer/easybill_java_sdk/resources/WebhookResource.java
@@ -1,7 +1,6 @@
 package io.github.bolzer.easybill_java_sdk.resources;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.github.bolzer.easybill_java_sdk.contracts.HttpClient;
 import io.github.bolzer.easybill_java_sdk.exceptions.EasybillRestException;
 import io.github.bolzer.easybill_java_sdk.models.Webhook;
@@ -17,7 +16,6 @@ public final class WebhookResource {
     public static final String RESOURCE_URL = "/webhooks";
 
     @NonNull
-    @SuppressFBWarnings
     private final HttpClient httpClient;
 
     public WebhookResource(@NonNull HttpClient httpClient) {
@@ -27,7 +25,7 @@ public final class WebhookResource {
     public @NonNull PaginatedResponse<Webhook> fetchWebhooks(
         GenericListQueryRequest genericListQueryRequest
     ) throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL,
                 genericListQueryRequest,
                 new TypeReference<>() {}
@@ -36,7 +34,7 @@ public final class WebhookResource {
 
     public @NonNull Webhook fetchWebhook(@Positive long webhookId)
         throws EasybillRestException {
-        return this.httpClient.getJson(
+        return this.httpClient.requestGetJson(
                 RESOURCE_URL + "/" + webhookId,
                 new TypeReference<>() {}
             );
@@ -45,7 +43,7 @@ public final class WebhookResource {
     public @NonNull Webhook createWebhook(
         @NonNull WebhookRequest webhookRequest
     ) throws EasybillRestException {
-        return this.httpClient.postJson(
+        return this.httpClient.requestPostJson(
                 RESOURCE_URL,
                 webhookRequest,
                 new TypeReference<>() {}
@@ -56,7 +54,7 @@ public final class WebhookResource {
         @Positive long webhookId,
         @NonNull WebhookRequest webhookRequest
     ) throws EasybillRestException {
-        return this.httpClient.putJson(
+        return this.httpClient.requestPutJson(
                 RESOURCE_URL + "/" + webhookId,
                 webhookRequest,
                 new TypeReference<>() {}
@@ -65,6 +63,6 @@ public final class WebhookResource {
 
     public void deleteWebhook(@Positive long webhookId)
         throws EasybillRestException {
-        this.httpClient.delete(RESOURCE_URL + "/" + webhookId);
+        this.httpClient.requestDelete(RESOURCE_URL + "/" + webhookId);
     }
 }